# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SessionLayer Control Plane API
  version: 0.1.0
  summary: Contract-first REST surface for the SessionLayer Control Plane.
  description: |
    The **SessionLayer Control Plane API** is the contract-first source of truth
    for the platform's REST surface (Design §13, FR-API-1). The UI is a *client* of
    this contract; server interfaces and typed clients are code-generated from this
    document, and CI fails if generated code drifts from the spec. Postgres is the
    single source of truth; the UI and the API share one write path (D11).

    **Complete contract.** This document is the complete resource surface: the
    config resources (`node-policies`, `rules`, `roles` + `role-bindings`, `cas`,
    `capability-defs`, `jit-policies`, `breakglass-policies`,
    `service-accounts`, `session-limit-policies`, `operator-settings`), the
    runtime resources (`nodes`, `gateways`, `sessions`, `session-leases`,
    `locks`, `jit-requests`, `join-tokens`, `otp`, device-flow, `pins`,
    `breakglass`, `recordings`, `audit-events`), and the operational probes.

    **Config vs runtime.** Config resources carry an `origin` provenance label
    (`api|ui|default`) and an optimistic-concurrency `version`; runtime resources
    (locks, sessions, issuance, audit) are never provenance-labelled. `locks` is an
    API-only runtime resource (Design §13).

    **Conventions.** URI versioning (`/v1`); cursor pagination on collections;
...
    compatibility policy.
...
servers:
...
security:
  - oidcBearer: []
  - clientCredentials: []
  - mtls: []
paths:
...
  /v1/role-bindings:
    get:
      operationId: listRoleBindings
      summary: List role bindings (admin).
      description: |
        Lists platform role bindings (`config.role_binding`; FR-PADM-2) — subject
        (user|group) -> role, optionally scoped for `recording:replay/export`.
        Cursor-paginated. Platform-RBAC gated (`rbac:read`).
      tags:
        - role-bindings
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of role bindings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleBindingPage'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createRoleBinding
      summary: Create a role binding (admin).
      description: |
        Binds a subject to a role. An unknown `roleId` or a malformed scope is a
        `422`; a duplicate (role, subjectKind, subject) is a `409`. Platform-RBAC
        gated (`rbac:write`) + audited.
      tags:
        - role-bindings
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleBindingRequest'
      responses:
        '201':
          description: The role binding was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleBindingResource'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/role-bindings/{bindingId}:
    parameters:
      - $ref: '#/components/parameters/BindingId'
    get:
      operationId: getRoleBinding
      summary: Get a role binding (admin).
      description: Returns one role binding by id. Platform-RBAC gated (`rbac:read`).
      tags:
        - role-bindings
      responses:
        '200':
          description: The role binding.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleBindingResource'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    put:
      operationId: updateRoleBinding
      summary: Update a role binding (admin).
      description: |
        Replaces a binding's scope (the subject and role are immutable — rebind by
        delete + create). A stale `version` is a `409`. Platform-RBAC gated
        (`rbac:write`) + audited.
      tags:
        - role-bindings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleBindingRequest'
      responses:
        '200':
          description: The updated role binding.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleBindingResource'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteRoleBinding
      summary: Delete a role binding (admin).
      description: Deletes a role binding (idempotent). Platform-RBAC gated (`rbac:write`) + audited.
      tags:
        - role-bindings
      responses:
        '204':
          description: The role binding was deleted (idempotent).
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
  /v1/sessions:
    get:
      operationId: listSessions
      summary: List SSH sessions (admin).
      description: |
        Lists SSH sessions (`runtime.ssh_session`; Design §12A) with their decision
        snapshot (identity, node, principal, access model, capabilities, grant
        expiry). Optional filters narrow by identity, node, access model, and
        active-only. Cursor-paginated. Platform-RBAC gated (`audit:read`).
      tags:
        - sessions
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: identity
          in: query
          required: false
          description: Filter to one subject identity.
          schema:
            type: string
        - name: nodeId
          in: query
          required: false
          description: Filter to one node id.
          schema:
            type: string
            format: uuid
        - name: accessModel
          in: query
          required: false
          description: Filter to one access model.
          schema:
            type: string
            enum: [standing, jit, breakglass]
        - name: activeOnly
          in: query
          required: false
          description: When true, return only sessions that have not ended.
          schema:
            type: boolean
      responses:
        '200':
          description: A page of sessions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionPage'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/sessions/{sessionId}:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      operationId: getSession
      summary: Get an SSH session (admin).
      description: Returns one session by id. Platform-RBAC gated (`audit:read`).
      tags:
        - sessions
      responses:
        '200':
          description: The session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResource'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/sessions/{sessionId}/terminate:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    post:
      operationId: terminateSession
      summary: Terminate an SSH session (admin).
      description: |
        Tears down a live session by pushing a top-tier `Lock` (Design §8.3/§8.4)
        scoped to the session's subject identity, reusing the teardown path — a
        deny that wins and is actively pushed to every Gateway (fail-closed). Because
        the wire `Lock` selector has no per-session facet, the teardown is
        identity-scoped (it also affects that identity's other live sessions); it is
        bounded by a short TTL so the identity may immediately reconnect under an
        unchanged policy. Idempotent. Platform-RBAC gated (`lock:write`) + audited.
      tags:
        - sessions
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminateSessionRequest'
      responses:
        '202':
          description: The teardown lock was pushed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResource'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC bearer access via an ID/JWT token validated by the Control Plane
        (Design §5.7, FR-AUTH-17). The ID token is the authentication proof.
    clientCredentials:
      type: oauth2
      description: |
        OAuth 2.0 client-credentials for machine consumers (Design §5.6,
        FR-AUTH-12). `private_key_jwt`/mTLS client authentication is preferred
        over static secrets.
      flows:
        clientCredentials:
          tokenUrl: https://controlplane.sessionlayer.example/oauth2/token
          scopes: {}
    mtls:
      type: mutualTLS
      description: |
        Mutual TLS client-certificate authentication (Design §5.7, FR-AUTH-17),
        also used by the internal CP<->Gateway plane.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: |
        Opaque forward-only pagination cursor from a previous page's `nextCursor`.
        Omit for the first page. An unrecognised cursor is a `400`.
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum items to return in one page (the server clamps to its maximum).
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 50
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Optional client-generated idempotency key scoping a mutating request. A
        retry with the same key, method, and path returns the original response and
        never repeats the side effect; the same key with a different request body is
        a `422`. Keys are retained for a bounded TTL.
      schema:
        type: string
        minLength: 1
        maxLength: 255
...
    BindingId:
      name: bindingId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Problem:
      description: An error occurred. The body is an RFC 9457 problem document.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      title: RFC 9457 Problem Details
      description: |
        Problem details for HTTP APIs as defined by RFC 9457 (obsoletes RFC 7807).
        Error responses use media type `application/problem+json`.
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
          description: A URI reference identifying the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: The HTTP status code.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: A URI reference identifying the specific occurrence.
      additionalProperties: true
...
    Origin:
      type: string
      title: Config Provenance
      description: |
        Provenance of a config row (Design §13). `api`/`ui` mark which admin surface
        last wrote it; `default` is a seeded/cold-start default. Read-only.
      enum:
        - api
        - ui
        - default
    Capability:
      type: string
      title: SSH Capability
      description: A policy-gated SSH channel capability (Design D14).
      enum:
        - shell
        - exec
        - sftp
        - scp
        - port_forward_local
        - port_forward_remote
        - agent_forward
        - x11
...
    SubjectKind:
      type: string
      title: Binding Subject Kind
      enum:
        - user
        - group
...
    AccessModel:
      type: string
      title: Access Model
      enum:
        - standing
        - jit
        - breakglass
...
    Selector:
      type: object
      title: Selector
      description: A shape-validated selector object (stored as jsonb).
      additionalProperties: true
...
    RoleBindingResource:
      type: object
      title: Role Binding
      required:
        - id
        - roleId
        - subjectKind
        - subject
        - origin
        - version
      properties:
        id:
          type: string
          format: uuid
        roleId:
          type: string
          format: uuid
        subjectKind:
          $ref: '#/components/schemas/SubjectKind'
        subject:
          type: string
        scope:
          $ref: '#/components/schemas/Selector'
        origin:
          $ref: '#/components/schemas/Origin'
        version:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    CreateRoleBindingRequest:
      type: object
      title: Create Role Binding Request
      required:
        - roleId
        - subjectKind
        - subject
      properties:
        roleId:
          type: string
          format: uuid
        subjectKind:
          $ref: '#/components/schemas/SubjectKind'
        subject:
          type: string
          minLength: 1
        scope:
          $ref: '#/components/schemas/Selector'
      additionalProperties: false
    UpdateRoleBindingRequest:
      type: object
      title: Update Role Binding Request
      required:
        - version
      properties:
        scope:
          $ref: '#/components/schemas/Selector'
        version:
          type: integer
          format: int64
      additionalProperties: false
    RoleBindingPage:
      type: object
      title: Role Binding Page
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RoleBindingResource'
        nextCursor:
          type: string
      additionalProperties: false
    # ============ cas (ca_config) ============
...
    SessionResource:
      type: object
      title: SSH Session
      required:
        - id
        - identity
        - principal
        - accessModel
        - capabilities
        - startedAt
      properties:
        id:
          type: string
          format: uuid
        identity:
          type: string
        nodeId:
          type: string
          format: uuid
        nodeName:
          type: string
        principal:
          type: string
        gatewayId:
          type: string
          format: uuid
        gatewayName:
          type: string
        accessModel:
          $ref: '#/components/schemas/AccessModel'
        capabilities:
          type: array
          items:
            $ref: '#/components/schemas/Capability'
        matchedRuleId:
          type: string
          format: uuid
        matchedRuleName:
          type: string
        jitRequestId:
          type: string
          format: uuid
        breakglassActivationId:
          type: string
          format: uuid
        policyEpoch:
          type: integer
          format: int64
        grantExpiry:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        endReason:
          type: string
      additionalProperties: false
    TerminateSessionRequest:
      type: object
      title: Terminate Session Request
      properties:
        reason:
          type: string
          maxLength: 4096
          description: |
            Operator/audit reason (never disclosed to the SSH user). Bounded like a
            Lock reason (it becomes one) so it cannot inflate the pushed deny-list.
      additionalProperties: false
    SessionPage:
      type: object
      title: Session Page
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SessionResource'
        nextCursor:
          type: string
      additionalProperties: false
    # ============ session-leases (session_lease, runtime) ============
...